applications/python-code/requirements.txt
setuptools
paho-mqtt==1.4.0
jsonschema>=3.0.2
flask>=1.1.1
requests>=2.22.0
gunicorn>=19.9.0
requests-mock>=1.7.0
SQLAlchemy>=1.3.8
psycopg2-binary>=2.8.3